Cards Can now be tapped and loads up all the descriptions for that ca…

+++ app/src/main/java/com/amcwustl/dailytarot/data/CardDbHelper.java
@@ -159,7 +159,7 @@ public class CardDbHelper extends SQLiteOpenHelper {
     }
   }
 
-  public Card getCardById(int cardId) {
+  public Card getCardById(Long cardId) {
     SQLiteDatabase db = getReadableDatabase();
     Card card = null;
